refactor(schedule-details): trim chart-controls PR helpers, tests, and mocks
Fold single-caller helpers (shift-chart-time-window, resolve-chart-span-
from-marker-interval-ms) into the modules that exclusively use them, drop
a redundant test case, and extract shared mock fixtures for the repeated
CadenceScheduleTime search-attribute encoding.
