+++ src/views/schedule-details/schedule-details-runs-chart/helpers/__tests__/chart-view-state.test.ts
@@ -7,6 +7,7 @@ import {
   isSameChartTimeWindow,
   panChartTimeWindowToTime,
   resolveChartFollowTimeWindow,
+  shiftChartTimeWindow,
   zoomChartTimeWindow,
 } from '../chart-view-state';
 
@@ -136,6 +137,37 @@ describe(panChartTimeWindowToTime.name, () => {
     });
 
     expect(panned.maxMs).toBe(bounds.maxMs);
+  });
+});
+
+describe(shiftChartTimeWindow.name, () => {
+  it('shifts both bounds by deltaMs when unbounded', () => {
+    expect(
+      shiftChartTimeWindow({
+        visibleWindow: { minMs: 10_000, maxMs: 20_000 },
+        deltaMs: -5_000,
+      })
+    ).toEqual({ minMs: 5_000, maxMs: 15_000 });
+  });
+
+  it('clamps the shift to the lower bound, preserving span', () => {
+    expect(
+      shiftChartTimeWindow({
+        visibleWindow: { minMs: 10_000, maxMs: 20_000 },
+        deltaMs: -50_000,
+        bounds: { minMs: 0, maxMs: 100_000 },
+      })
+    ).toEqual({ minMs: 0, maxMs: 10_000 });
+  });
+
+  it('clamps the shift to the upper bound, preserving span', () => {
+    expect(
+      shiftChartTimeWindow({
+        visibleWindow: { minMs: 10_000, maxMs: 20_000 },
+        deltaMs: 500_000,
+        bounds: { minMs: 0, maxMs: 100_000 },
+      })
+    ).toEqual({ minMs: 90_000, maxMs: 100_000 });
   });
 });
