+++ src/views/schedule-details/schedule-details-runs-chart/helpers/chart-view-state.ts
@@ -7,6 +7,7 @@ import {
   type ChartTimeWindow,
   type PanChartTimeWindowToTimeParams,
   type ResolveChartFollowTimeWindowParams,
+  type ShiftChartTimeWindowParams,
   type ZoomChartTimeWindowParams,
 } from '../schedule-details-runs-chart.types';
 
@@ -142,6 +143,30 @@ export function resolveChartFollowTimeWindow({
   return nextRunAnchoredWindow.minMs <= nowMs && spanMs > 0
     ? nextRunAnchoredWindow
     : nowAnchoredWindow;
+}
+
+export function shiftChartTimeWindow({
+  visibleWindow,
+  deltaMs,
+  bounds,
+}: ShiftChartTimeWindowParams): ChartTimeWindow {
+  const spanMs = getChartTimeWindowSpanMs(visibleWindow);
+  let minMs = visibleWindow.minMs + deltaMs;
+  let maxMs = visibleWindow.maxMs + deltaMs;
+
+  if (bounds) {
+    if (minMs < bounds.minMs) {
+      minMs = bounds.minMs;
+      maxMs = minMs + spanMs;
+    }
+
+    if (maxMs > bounds.maxMs) {
+      maxMs = bounds.maxMs;
+      minMs = maxMs - spanMs;
+    }
+  }
+
+  return { minMs, maxMs };
 }
 
 export function canZoomChartIn(visibleWindow: ChartTimeWindow): boolean {
